refactor(icons): move useScalarIcon to hooks directory
