+++ packages/icons/src/hooks/useScalarIcon.ts
