+++ packages/icons/src/scripts/generate-icons.ts
@@ -117,7 +117,7 @@ export default {
 </script>
 <script lang="ts" setup>
 import { computed } from 'vue'
-import { type ScalarIconProps, useScalarIcon } from '../useScalarIcon'
+import { type ScalarIconProps, useScalarIcon } from '../hooks/useScalarIcon'
 
 const props = defineProps<ScalarIconProps>()
